migration/pyproject.toml
[tool.poetry]
name = "migration"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jira = "^3.8.0"
pandas = "^2.2.2"
pygithub = "^2.3.0"
termcolor = "^2.4.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
